The problems with removing rows or columns in PHPExcel/PhpSpreadsheet are more extensive than I recognized with my previous post/patch. For example, removing all of the columns associated with a merged range of cells results in there continuing to be a merged range of cells that doesn't belong. I've updated the ReferenceHelper to return an empty string for the updated range if the range has been eliminated by a removal and updated the callers to the best of my understanding. For example, in the case of merged ranges, receiving an empty string means the merged range should be discarded. In the case of recomputing the freezePane, and empty string is the perfect value to pass to the worksheet.
I also recognized why an "if (...) { continue; }" was present that I
removed in my previous patch. That "if" block does cause problems for
the leading columns in the case of a removal, but it should be fixed,
not removed. I've updated it to only skip the columns that are being
removed.